+++ tests-corpus/valid-bare-schema/list-union-circular-ref/out.gen.d.ts
@@ -1,7 +1,7 @@
 import * as bare from "@bare-ts/lib"
 
 export type Alias = readonly (
-    | { readonly tag: 0, readonly val: Alias }
+    | { readonly tag: "Alias", readonly val: Alias }
     | { readonly tag: 1, readonly val: string })[]
 
 export function readAlias(bc: bare.ByteCursor): Alias