+++ tests-corpus/valid-bare-schema/optional-union-circular/out.gen.ts
@@ -1,13 +1,13 @@
 import * as bare from "@bare-ts/lib"
 
 function read0(bc: bare.ByteCursor): 
-    | { readonly tag: 0, readonly val: Alias }
+    | { readonly tag: "Alias", readonly val: Alias }
     | { readonly tag: 1, readonly val: string } {
     const offset = bc.offset
     const tag = bare.readU8(bc)
     switch (tag) {
         case 0:
-            return { tag, val: readAlias(bc) }
+            return { tag: "Alias", val: readAlias(bc) }
         case 1:
             return { tag, val: bare.readString(bc) }
         default: {
@@ -18,11 +18,11 @@ function read0(bc: bare.ByteCursor):
 }
 
 function write0(bc: bare.ByteCursor, x: 
-    | { readonly tag: 0, readonly val: Alias }
+    | { readonly tag: "Alias", readonly val: Alias }
     | { readonly tag: 1, readonly val: string }): void {
-    bare.writeU8(bc, x.tag)
     switch (x.tag) {
-        case 0: {
+        case "Alias": {
+            bare.writeU8(bc, 0)
             writeAlias(bc, x.val)
             break
         }
@@ -34,7 +34,7 @@ function write0(bc: bare.ByteCursor, x:
 }
 
 export type Alias =
-    | { readonly tag: 0, readonly val: Alias }
+    | { readonly tag: "Alias", readonly val: Alias }
     | { readonly tag: 1, readonly val: string } | null
 
 export function readAlias(bc: bare.ByteCursor): Alias {