+++ tests-corpus/valid-bare-schema/struct-circular-ref-allowed/out.gen.d.ts
@@ -3,7 +3,7 @@ import * as bare from "@bare-ts/lib"
 export type Person = {
     readonly bestFriend: Person | null,
     readonly secondBestFriend: 
-        | { readonly tag: 0, readonly val: Person }
+        | { readonly tag: "Person", readonly val: Person }
         | { readonly tag: 1, readonly val: null },
     readonly friends: readonly Person[],
     readonly friendNicknames: ReadonlyMap<string, Person>,