+++ tests-corpus/valid-bare-schema/union-circular/out.gen.ts
@@ -3,17 +3,17 @@ import * as bare from "@bare-ts/lib"
 export type u8 = number
 
 export type Alias =
-    | { readonly tag: 0, readonly val: Alias1 }
-    | { readonly tag: 1, readonly val: Alias2 }
+    | { readonly tag: "Alias1", readonly val: Alias1 }
+    | { readonly tag: "Alias2", readonly val: Alias2 }
 
 export function readAlias(bc: bare.ByteCursor): Alias {
     const offset = bc.offset
     const tag = bare.readU8(bc)
     switch (tag) {
         case 0:
-            return { tag, val: readAlias1(bc) }
+            return { tag: "Alias1", val: readAlias1(bc) }
         case 1:
-            return { tag, val: readAlias2(bc) }
+            return { tag: "Alias2", val: readAlias2(bc) }
         default: {
             bc.offset = offset
             throw new bare.BareError(offset, "invalid tag")
@@ -22,13 +22,14 @@ export function readAlias(bc: bare.ByteCursor): Alias {
 }
 
 export function writeAlias(bc: bare.ByteCursor, x: Alias): void {
-    bare.writeU8(bc, x.tag)
     switch (x.tag) {
-        case 0: {
+        case "Alias1": {
+            bare.writeU8(bc, 0)
             writeAlias1(bc, x.val)
             break
         }
-        case 1: {
+        case "Alias2": {
+            bare.writeU8(bc, 1)
             writeAlias2(bc, x.val)
             break
         }