+++ tests-corpus/valid-bare-schema/union-struct-int-tag/out.gen.js
@@ -2,30 +2,34 @@ import * as bare from "@bare-ts/lib"
 
 const config = /* @__PURE__ */ bare.Config({})
 
-export type u8 = number
-
-export type Y = u8
-
-export function readY(bc: bare.ByteCursor): Y {
-    return bare.readU8(bc)
+export function readBoxedU32(bc) {
+    return {
+        val: bare.readU32(bc),
+    }
 }
 
-export function writeY(bc: bare.ByteCursor, x: Y): void {
-    bare.writeU8(bc, x)
+export function writeBoxedU32(bc, x) {
+    bare.writeU32(bc, x.val)
 }
 
-export type X =
-    | { readonly tag: 0, readonly val: u8 }
-    | { readonly tag: "Y", readonly val: Y }
+export function readBoxedStr(bc) {
+    return {
+        val: bare.readString(bc),
+    }
+}
 
-export function readX(bc: bare.ByteCursor): X {
+export function writeBoxedStr(bc, x) {
+    bare.writeString(bc, x.val)
+}
+
+export function readBoxed(bc) {
     const offset = bc.offset
     const tag = bare.readU8(bc)
     switch (tag) {
         case 0:
-            return { tag, val: bare.readU8(bc) }
+            return { tag, val: readBoxedU32(bc) }
         case 1:
-            return { tag: "Y", val: readY(bc) }
+            return { tag, val: readBoxedStr(bc) }
         default: {
             bc.offset = offset
             throw new bare.BareError(offset, "invalid tag")
@@ -33,32 +37,32 @@ export function readX(bc: bare.ByteCursor): X {
     }
 }
 
-export function writeX(bc: bare.ByteCursor, x: X): void {
+export function writeBoxed(bc, x) {
+    bare.writeU8(bc, x.tag)
     switch (x.tag) {
         case 0: {
-            bare.writeU8(bc, x.val)
+            writeBoxedU32(bc, x.val)
             break
         }
-        case "Y": {
-            bare.writeU8(bc, 1)
-            writeY(bc, x.val)
+        case 1: {
+            writeBoxedStr(bc, x.val)
             break
         }
     }
 }
 
-export function encodeX(x: X): Uint8Array {
+export function encodeBoxed(x) {
     const bc = new bare.ByteCursor(
         new Uint8Array(config.initialBufferLength),
         config
     )
-    writeX(bc, x)
+    writeBoxed(bc, x)
     return new Uint8Array(bc.view.buffer, bc.view.byteOffset, bc.offset)
 }
 
-export function decodeX(bytes: Uint8Array): X {
+export function decodeBoxed(bytes) {
     const bc = new bare.ByteCursor(bytes, config)
-    const result = readX(bc)
+    const result = readBoxed(bc)
     if (bc.offset < bc.view.byteLength) {
         throw new bare.BareError(bc.offset, "remaining bytes")
     }