+++ tests-corpus/valid-bare-schema/x-bareorg-example/out.gen.ts
@@ -231,20 +231,20 @@ export function writeEmployee(bc: bare.ByteCursor, x: Employee): void {
 export type TerminatedEmployee = null
 
 export type Person =
-    | { readonly tag: 0, readonly val: Customer }
-    | { readonly tag: 1, readonly val: Employee }
-    | { readonly tag: 2, readonly val: TerminatedEmployee }
+    | { readonly tag: "Customer", readonly val: Customer }
+    | { readonly tag: "Employee", readonly val: Employee }
+    | { readonly tag: "TerminatedEmployee", readonly val: TerminatedEmployee }
 
 export function readPerson(bc: bare.ByteCursor): Person {
     const offset = bc.offset
     const tag = bare.readU8(bc)
     switch (tag) {
         case 0:
-            return { tag, val: readCustomer(bc) }
+            return { tag: "Customer", val: readCustomer(bc) }
         case 1:
-            return { tag, val: readEmployee(bc) }
+            return { tag: "Employee", val: readEmployee(bc) }
         case 2:
-            return { tag, val: null }
+            return { tag: "TerminatedEmployee", val: null }
         default: {
             bc.offset = offset
             throw new bare.BareError(offset, "invalid tag")
@@ -253,14 +253,19 @@ export function readPerson(bc: bare.ByteCursor): Person {
 }
 
 export function writePerson(bc: bare.ByteCursor, x: Person): void {
-    bare.writeU8(bc, x.tag)
     switch (x.tag) {
-        case 0: {
+        case "Customer": {
+            bare.writeU8(bc, 0)
             writeCustomer(bc, x.val)
             break
         }
-        case 1: {
+        case "Employee": {
+            bare.writeU8(bc, 1)
             writeEmployee(bc, x.val)
+            break
+        }
+        case "TerminatedEmployee": {
+            bare.writeU8(bc, 2)
             break
         }
     }