+++ tests-corpus/valid-bare-schema/x-readme-example/out.gen.js
@@ -86,9 +86,9 @@ export function readContact(bc) {
     const tag = bare.readU8(bc)
     switch (tag) {
         case 0:
-            return { tag, val: readPerson(bc) }
+            return { tag: "Person", val: readPerson(bc) }
         case 1:
-            return { tag, val: readOrganization(bc) }
+            return { tag: "Organization", val: readOrganization(bc) }
         default: {
             bc.offset = offset
             throw new bare.BareError(offset, "invalid tag")
@@ -97,13 +97,14 @@ export function readContact(bc) {
 }
 
 export function writeContact(bc, x) {
-    bare.writeU8(bc, x.tag)
     switch (x.tag) {
-        case 0: {
+        case "Person": {
+            bare.writeU8(bc, 0)
             writePerson(bc, x.val)
             break
         }
-        case 1: {
+        case "Organization": {
+            bare.writeU8(bc, 1)
             writeOrganization(bc, x.val)
             break
         }